pyproject.toml
[project]
name = "ai-organizational-changes"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Marc Fabian Mezger", email = "57255687+mfmezger@users.noreply.github.com" }
]
requires-python = ">=3.13"
dependencies = [
    "loguru>=0.7.3",
    "openpyxl>=3.1.5",
    "pandas>=2.3.2",
    "pydantic-ai>=1.0.10",
    "python-dotenv>=1.1.1",
]

[project.scripts]
ai-organizational-changes = "ai_organizational_changes:main"

[build-system]
requires = ["uv_build>=0.8.22,<0.9.0"]
build-backend = "uv_build"
